+++ src/main.rs
@@ -1,6 +1,9 @@
+mod error;
 mod ssr;
 
-use crate::ssr::{cli::Cli, error::Result, retriever::SsrRetriever};
+use crate::error::Result;
+use crate::ssr::Cli;
+use crate::ssr::SsrRetriever;
 
 fn main() -> Result<()> {
     let cli = Cli::parse_args();
