+++ src/ssr.rs
@@ -1,4 +1,6 @@
-pub mod cli;
-pub mod error;
-pub mod retriever;
-mod types;
+mod cli;
+mod retriever;
+mod ssr;
+
+pub use self::cli::*;
+pub use self::retriever::*;
